restrictions to some endpoints. security improvements

+++ ScienceArchive.Server/ScienceArchive.Web.Api/Controllers/ArticleController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using ScienceArchive.Application.Dtos.Article.Request;
 using ScienceArchive.Application.Interfaces.Services;
 using ScienceArchive.Web.Api.Auth;
@@ -18,18 +19,19 @@ public class ArticleController : ControllerBase
     }
     
     [HttpGet("by-category/{categoryId}")]
-    public async Task<Response> GetByCategoryId(string categoryId)
+    public async Task<Response> GetVerifiedByCategoryId(string categoryId)
     {
         if (string.IsNullOrWhiteSpace(categoryId))
         {
             throw new BadHttpRequestException("Category ID was not presented");
         }
         
-        var dto = new GetArticlesByCategoryIdRequestDto(categoryId);
-        var result = await _articleService.GetArticlesByCategoryId(dto);
+        var dto = new GetVerifiedArticlesByCategoryIdRequestDto(categoryId);
+        var result = await _articleService.GetVerifiedArticlesByCategoryId(dto);
         return new SuccessResponse(result);
     }
     
+    [Authorize]
     [HttpGet("my-articles")]
     public async Task<Response> GetUserArticles()
     {
@@ -97,6 +99,7 @@ public class ArticleController : ControllerBase
         return new SuccessResponse(result);
     }
     
+    [Authorize]
     [HttpPost("create")]
     public async Task<Response> Create([FromBody] CreateArticleRequestDto? dto)
     {
@@ -105,10 +108,19 @@ public class ArticleController : ControllerBase
             throw new BadHttpRequestException("No data presented");
         }
         
-        var result = await _articleService.CreateArticle(dto);
+        var userId = HttpContext.GetUserIdFromToken();
+
+        if (string.IsNullOrWhiteSpace(userId))
+        {
+            throw new BadHttpRequestException("Cannot get user ID", 403);
+        }
+
+        var populatedDto = dto with { UserId = userId };
+        var result = await _articleService.CreateArticle(populatedDto);
         return new SuccessResponse(result);
     }
     
+    [Authorize]
     [HttpPost("update")]
     public async Task<Response> Update([FromBody] UpdateArticleRequestDto? dto)
     {
@@ -117,10 +129,19 @@ public class ArticleController : ControllerBase
             throw new BadHttpRequestException("No data presented");
         }
         
-        var result = await _articleService.UpdateArticle(dto);
+        var userId = HttpContext.GetUserIdFromToken();
+
+        if (string.IsNullOrWhiteSpace(userId))
+        {
+            throw new BadHttpRequestException("Cannot get user ID", 403);
+        }
+
+        var populatedDto = dto with { UserId = userId};
+        var result = await _articleService.UpdateArticle(populatedDto);
         return new SuccessResponse(result);
     }
 
+    [AuthorizeClaims(AuthClaims.ApproveArticles)]
     [HttpPost("approve")]
     public async Task<Response> Approve([FromBody] ApproveArticleRequestDto? dto)
     {
@@ -133,7 +154,7 @@ public class ArticleController : ControllerBase
         return new SuccessResponse(result);
     }
     
-    
+    [AuthorizeClaims(AuthClaims.DeclineArticles)]
     [HttpPost("decline")]
     public async Task<Response> Decline([FromBody] DeclineArticleRequestDto? dto)
     {
@@ -146,6 +167,7 @@ public class ArticleController : ControllerBase
         return new SuccessResponse(result);
     }
     
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<Response> Delete(string? id)
     {
@@ -154,7 +176,14 @@ public class ArticleController : ControllerBase
             throw new BadHttpRequestException("No data presented");
         }
         
-        var dto = new DeleteArticleRequestDto(id);
+        var userId = HttpContext.GetUserIdFromToken();
+
+        if (string.IsNullOrWhiteSpace(userId))
+        {
+            throw new BadHttpRequestException("Cannot get user ID", 403);
+        }
+        
+        var dto = new DeleteArticleRequestDto(id, userId);
         var result = await _articleService.DeleteArticle(dto);
         return new SuccessResponse(result);
     }
